docs / proper code splitting
